requirements.txt
browser-use==0.8.0
python-dotenv==1.1.1
openai==1.109.1
playwright==1.55.0
